rename blockdb package to persist package
