+++ modules/consensus/persist.go
@@ -3,12 +3,12 @@ package consensus
 import (
 	"path/filepath"
 
-	"github.com/NebulousLabs/Sia/blockdb"
+	"github.com/NebulousLabs/Sia/persist"
 	"github.com/NebulousLabs/Sia/types"
 )
 
 func (s *State) load(saveDir string) error {
-	db, err := blockdb.Open(filepath.Join(saveDir, "chain.db"))
+	db, err := persist.OpenDB(filepath.Join(saveDir, "chain.db"))
 	if err != nil {
 		return err
 	}
@@ -24,7 +24,7 @@ func (s *State) load(saveDir string) error {
 	// load blocks from the db, starting after the genesis block
 	// NOTE: during load, the state uses the NilDB. This prevents AcceptBlock
 	// from adding duplicate blocks to the real database.
-	s.db = blockdb.NilDB
+	s.db = persist.NilDB
 	for i := types.BlockHeight(1); i < height; i++ {
 		b, err := db.Block(i)
 		if err != nil {
